remove pyi files
